+++ tools/oracle/engine/backends/scapy/packets.py
@@ -35,6 +35,19 @@ from .encode_helpers import (
 # are empty and every stack/layer falls through to the legacy code.
 from .protocols import SCAPY_REGISTRY, STACK_ENCODER_REGISTRY
 
+# The IPv6 extension-header builders live in the co-located ipv6 plugin module
+# (``protocols/ipv6.py``) but, unlike the base ``ipv6`` layer, are not registered in
+# ``SCAPY_REGISTRY`` because the ext-header names are sub-layers of the ``ipv6`` spec
+# rather than top-level spec layers. They are re-imported here so the legacy per-layer
+# ``_build_layer`` dispatch keeps routing ``ipv6_hop_by_hop`` /
+# ``ipv6_destination_options`` / ``ipv6_fragment`` / ``ipv6_routing`` to them unchanged.
+from .protocols.ipv6 import (
+    _ipv6_destination_options,
+    _ipv6_fragment,
+    _ipv6_hop_by_hop,
+    _ipv6_routing,
+)
+
 
 BACKEND_NAME = "scapy"
 
@@ -1241,189 +1254,6 @@ def _ospf_lsa_list(value: object, scapy_ospf: Any) -> list[Any]:
             else:
                 lsas.append(header)
     return lsas
-
-
-def _ipv6_hop_by_hop(
-    fields: Mapping[str, JSONObject],
-    stack: list[str],
-    index: int,
-    scapy_all: Any,
-) -> Any:
-    return _ipv6_options_header(
-        fields,
-        layer="ipv6_hop_by_hop",
-        factory=scapy_all.IPv6ExtHdrHopByHop,
-        scapy_all=scapy_all,
-    )
-
-
-def _ipv6_destination_options(
-    fields: Mapping[str, JSONObject],
-    stack: list[str],
-    index: int,
-    scapy_all: Any,
-) -> Any:
-    return _ipv6_options_header(
-        fields,
-        layer="ipv6_destination_options",
-        factory=scapy_all.IPv6ExtHdrDestOpt,
-        scapy_all=scapy_all,
-    )
-
-
-def _ipv6_options_header(
-    fields: Mapping[str, JSONObject],
-    *,
-    layer: str,
-    factory: Any,
-    scapy_all: Any,
-) -> Any:
-    option_fields = _layer_fields(fields, layer)
-    kwargs: dict[str, Any] = {
-        "nh": _protocol_value(
-            _required_field(option_fields, layer, "next_header", "nh"),
-            _IPV6_NEXT_HEADERS,
-        ),
-        # Scapy aligns some known options by inserting PadN ahead of them. The
-        # oracle model is byte-preserving, so plans carry any required padding
-        # explicitly and Scapy's alignment autopad must stay disabled.
-        "autopad": 0,
-        "options": _ipv6_options(_required_field(option_fields, layer, "options"), scapy_all),
-    }
-    if "header_ext_len" in option_fields or "len" in option_fields:
-        kwargs["len"] = _int(_optional_field(option_fields, "header_ext_len", "len"), 0)
-    return factory(**kwargs)
-
-
-def _ipv6_options(value: object, scapy_all: Any) -> list[Any]:
-    if not isinstance(value, Sequence) or isinstance(value, (str, bytes, bytearray)):
-        raise ValueError("IPv6 options materialization requires an option list")
-    return [_ipv6_option(item, scapy_all) for item in value]
-
-
-def _ipv6_option(item: object, scapy_all: Any) -> Any:
-    if not isinstance(item, Mapping):
-        raise ValueError(f"IPv6 option entry must be an object, got {item!r}")
-    kind = _ipv6_option_kind(item)
-    if kind == "pad1":
-        return scapy_all.Pad1()
-    if kind == "padn":
-        return scapy_all.PadN(optdata=_ipv6_padn_data(item))
-    if kind == "router_alert":
-        return scapy_all.RouterAlert(value=_int(_required_field(item, "ipv6 option", "value"), 0))
-    if kind == "jumbo_payload":
-        length = _required_field(item, "ipv6 option", "length", "jumbo_payload_length")
-        return scapy_all.Jumbo(jumboplen=_int(length, 0))
-    if kind == "home_address":
-        address = _required_field(item, "ipv6 option", "address", "home_address")
-        return scapy_all.HAO(hoa=_text(address, "::"))
-    option_type = _int(
-        _required_field(item, "ipv6 option", "option_type", "type", "kind"),
-        0,
-    )
-    return scapy_all.HBHOptUnknown(
-        otype=option_type,
-        optdata=_ipv6_option_data(item),
-    )
-
-
-def _ipv6_option_kind(item: Mapping[str, object]) -> str:
-    value = _optional_field(item, "kind", "name")
-    if isinstance(value, str):
-        normalized = value.lower().replace("-", "_")
-        if normalized in {
-            "pad1",
-            "padn",
-            "router_alert",
-            "jumbo_payload",
-            "home_address",
-            "unknown",
-            "generic",
-        }:
-            return normalized
-    option_type = _optional_field(item, "option_type", "type")
-    if option_type is None:
-        return "unknown"
-    option_type_int = _int(option_type, 0)
-    if option_type_int == 0:
-        return "pad1"
-    if option_type_int == 1:
-        return "padn"
-    if option_type_int == 5 and "value" in item:
-        return "router_alert"
-    if option_type_int == 0xC2 and ("length" in item or "jumbo_payload_length" in item):
-        return "jumbo_payload"
-    if option_type_int == 0xC9 and ("address" in item or "home_address" in item):
-        return "home_address"
-    return "unknown"
-
-
-def _ipv6_padn_data(item: Mapping[str, object]) -> bytes:
-    data = _ipv6_option_data(item)
-    if data:
-        return data
-    total_length = _optional_field(item, "total_length", "length")
-    if total_length is None:
-        return b""
-    total = _int(total_length, 0)
-    if total < 2:
-        raise ValueError(f"PadN total length must be at least 2 bytes, got {total}")
-    return b"\x00" * (total - 2)
-
-
-def _ipv6_option_data(item: Mapping[str, object]) -> bytes:
-    data = _optional_field(item, "data", "bytes", "value_hex", "hex")
-    if data is None:
-        return b""
-    return _bytes_field(data)
-
-
-def _ipv6_fragment(
-    fields: Mapping[str, JSONObject],
-    stack: list[str],
-    index: int,
-    scapy_all: Any,
-) -> Any:
-    fragment_fields = _layer_fields(fields, "ipv6_fragment")
-    kwargs: dict[str, Any] = {
-        "nh": _protocol_value(
-            _required_field(fragment_fields, "ipv6_fragment", "next_header", "nh"),
-            _IPV6_NEXT_HEADERS,
-        ),
-        "id": _int(_required_field(fragment_fields, "ipv6_fragment", "identification", "id"), 0),
-        "offset": _int(
-            _required_field(fragment_fields, "ipv6_fragment", "fragment_offset", "offset"),
-            0,
-        ),
-        "m": _int(_required_field(fragment_fields, "ipv6_fragment", "more_fragments", "m"), 0),
-    }
-    if "reserved" in fragment_fields:
-        kwargs["res1"] = _int(fragment_fields.get("reserved"), 0)
-    return scapy_all.IPv6ExtHdrFragment(**kwargs)
-
-
-def _ipv6_routing(
-    fields: Mapping[str, JSONObject],
-    stack: list[str],
-    index: int,
-    scapy_all: Any,
-) -> Any:
-    routing_fields = _layer_fields(fields, "ipv6_routing")
-    kwargs: dict[str, Any] = {
-        "nh": _protocol_value(
-            _required_field(routing_fields, "ipv6_routing", "next_header", "nh"),
-            _IPV6_NEXT_HEADERS,
-        ),
-        "type": _int(_required_field(routing_fields, "ipv6_routing", "type", "routing_type"), 0),
-        "segleft": _int(
-            _required_field(routing_fields, "ipv6_routing", "segments_left", "segleft"),
-            0,
-        ),
-    }
-    addresses = routing_fields.get("addresses")
-    if isinstance(addresses, list):
-        kwargs["addresses"] = addresses
-    return scapy_all.IPv6ExtHdrRouting(**kwargs)
 
 
 # ICMPv4 types whose rest-of-header Scapy's ICMP layer exposes via the id/seq
